Added 'directions.aiml'
This file lets users ask for directions and information about various
buildings.

The way I've done it allows us to add more locations at the top of the
file which should work straight away because of the general cases below.

I changed the 'bot.py' file so it doesn't save or load the brain because
this is pretty pointless atm :palm_tree: